one CB for both admin and tile overlays
